+++ state/graphql-sender/src/test/java/com/expediagroup/streamplatform/streamregistry/state/graphql/DefaultApolloClientFactoryTest.java
@@ -21,15 +21,14 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.util.function.Consumer;
+import okhttp3.OkHttpClient;
 
 import com.apollographql.apollo.ApolloClient;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
@@ -39,19 +38,15 @@ import org.mockito.junit.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class DefaultApolloClientFactoryTest {
   @Mock private ApolloClient.Builder builder;
-  @Mock private Consumer<ApolloClient.Builder> configurer;
 
   private final String streamRegistryUrl = "streamRegistryUrl";
 
   private DefaultApolloClientFactory underTest;
 
-  @Before
-  public void before() {
-    underTest = spy(new DefaultApolloClientFactory(streamRegistryUrl, configurer));
-  }
-
   @Test
-  public void test() {
+  public void test_withoutCredentials() {
+    underTest = spy(new DefaultApolloClientFactory(streamRegistryUrl));
+
     when(underTest.builder()).thenReturn(builder);
     when(builder.okHttpClient(any())).thenReturn(builder);
     when(builder.serverUrl(streamRegistryUrl)).thenReturn(builder);
@@ -60,13 +55,37 @@ public class DefaultApolloClientFactoryTest {
     underTest.create();
 
     var captor = ArgumentCaptor.forClass(ObjectNodeTypeAdapter.class);
-    var inOrder = inOrder(builder, configurer);
-    inOrder.verify(builder).okHttpClient(any());
-    inOrder.verify(configurer).accept(builder);
-    inOrder.verify(builder).serverUrl(streamRegistryUrl);
-    inOrder.verify(builder).addCustomTypeAdapter(eq(OBJECTNODE), captor.capture());
-    inOrder.verify(builder).build();
+    var okHttpClientCaptor = ArgumentCaptor.forClass(OkHttpClient.class);
 
+    verify(builder).okHttpClient(okHttpClientCaptor.capture());
+    verify(builder).serverUrl(streamRegistryUrl);
+    verify(builder).addCustomTypeAdapter(eq(OBJECTNODE), captor.capture());
+    verify(builder).build();
+
+    assertThat(okHttpClientCaptor.getValue().interceptors().size(), is(0));
     assertThat(captor.getValue(), is(notNullValue()));
+  }
+
+  @Test
+  public void test_withCredentials() {
+    underTest = spy(new DefaultApolloClientFactory(streamRegistryUrl, new Credentials("userName", "password")));
+
+    when(underTest.builder()).thenReturn(builder);
+    when(builder.okHttpClient(any())).thenReturn(builder);
+    when(builder.serverUrl(streamRegistryUrl)).thenReturn(builder);
+    when(builder.addCustomTypeAdapter(any(), any())).thenReturn(builder);
+
+    underTest.create();
+
+    var objectNodeTypeAdapterCaptor = ArgumentCaptor.forClass(ObjectNodeTypeAdapter.class);
+    var okHttpClientCaptor = ArgumentCaptor.forClass(OkHttpClient.class);
+
+    verify(builder).okHttpClient(okHttpClientCaptor.capture());
+    verify(builder).serverUrl(streamRegistryUrl);
+    verify(builder).addCustomTypeAdapter(eq(OBJECTNODE), objectNodeTypeAdapterCaptor.capture());
+    verify(builder).build();
+
+    assertThat(okHttpClientCaptor.getValue().interceptors().size(), is(1));
+    assertThat(objectNodeTypeAdapterCaptor.getValue(), is(notNullValue()));
   }
 }